Fix: AuthProvider context error
The error "useAuth must be used within an AuthProvider" indicates that the `useAuth` hook is being called outside of the `AuthProvider`. This commit will ensure that the `AuthProvider` wraps the components that use `useAuth`, specifically the `Login` component. This will be done by checking the `App.tsx` file and making sure that the `AuthProvider` is wrapping the `Routes` component.